# openapi.yaml
openapi: 3.0.3
info:
  title: nself-chat (nchat) API
  version: 0.2.0
  description: |
    **nself-chat (nchat)** is a white-label team communication platform combining features from Slack, Discord, and Telegram.

    This API provides endpoints for:
    - Application health and monitoring
    - Configuration management
    - Authentication
    - File uploads
    - Search
    - Webhooks
    - Audit logging
    - Translation and link previews

    ## Authentication

    Most endpoints require authentication via Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt-token>
    ```

    Admin-only endpoints require the authenticated user to have `owner` or `admin` role.

    ## Rate Limiting

    Most endpoints are rate-limited to prevent abuse:
    - Default: 60 requests per minute
    - Upload: 30 requests per minute
    - Search: 60 requests per minute
    - Webhooks: 30 requests per minute per webhook

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the limit resets

    ## Error Responses

    All error responses follow a consistent format:
    ```json
    {
      "success": false,
      "error": "Human-readable error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  contact:
    name: nself-chat Support
    url: https://github.com/nself-chat
  license:
...
servers:
...
paths:
...
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Create new user account
      description: Registers a new user account. First user becomes the owner.
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - username
              properties:
                email:
                  type: string
                  format: email
                  example: newuser@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securePassword123
                username:
                  type: string
                  pattern: '^[a-zA-Z0-9_-]{3,20}$'
                  example: johndoe
                displayName:
                  type: string
                  example: John Doe
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/translate:
    post:
      tags:
        - Translation
      summary: Translate text
      description: |
        Translates text to the target language. This is a stub implementation
        that can be integrated with services like Google Translate or DeepL.
      operationId: translateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - targetLanguage
              properties:
                text:
                  type: string
                  maxLength: 5000
                  example: Hello, how are you?
                targetLanguage:
                  type: string
                  example: es
                  description: ISO 639-1 language code
                sourceLanguage:
                  type: string
                  example: en
                  description: Source language (optional, will be detected)
      responses:
        '200':
          description: Translation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  translatedText:
                    type: string
                    example: Hola, ¿cómo estás?
                  sourceLanguage:
                    type: string
                    example: en
                  targetLanguage:
                    type: string
                    example: es
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/link-preview:
    get:
      tags:
        - Link Preview
      summary: Get link preview metadata
      description: Fetches OpenGraph and Twitter Card metadata for a URL
      operationId: getLinkPreview
      parameters:
        - name: url
          in: query
          required: true
          description: URL to fetch metadata for
          schema:
            type: string
            format: uri
            example: https://github.com
        - name: skipCache
          in: query
          description: Skip cache and fetch fresh data
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Preview metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkPreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/signin
  schemas:
...
    AuthResponse:
      type: object
      required:
        - user
        - accessToken
        - refreshToken
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          description: JWT access token (24h expiry)
        refreshToken:
          type: string
          description: JWT refresh token (30d expiry)
    User:
      type: object
      required:
        - id
        - email
        - username
        - role
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        role:
          type: string
          enum: [owner, admin, moderator, member, guest]
        status:
          type: string
          enum: [online, offline, away, busy]
...
    LinkPreviewResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        siteName:
          type: string
        cached:
          type: boolean
        error:
          type: string
        errorCode:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Invalid request
              code:
                type: string
                example: BAD_REQUEST
...