# src/app/api/translate/route.ts
/**
 * Translation API Route
 *
 * Translates text to target language using a translation service.
 * This is a stub that can be integrated with services like Google Translate, DeepL, etc.
 */

import { NextRequest, NextResponse } from 'next/server'
import { logger } from '@/lib/logger'

interface TranslationRequest {
  text: string
  targetLanguage: string
  sourceLanguage?: string
}

/**
 * Detect the language of the text (stub implementation)
 */
function detectLanguage(_text: string): string {
  // In production, use a proper language detection library
  // For now, return 'unknown'
  return 'unknown'
}

/**
 * Translate text (stub implementation)
 * In production, integrate with a translation API like:
 * - Google Cloud Translation API
 * - DeepL API
 * - Microsoft Translator
 * - AWS Translate
 */
async function translateText(
  text: string,
  targetLanguage: string,
  sourceLanguage?: string
): Promise<{ translatedText: string; sourceLanguage: string }> {
  // This is a stub implementation
  // In production, replace with actual API calls

  // For demo purposes, just add a prefix
  const detectedSource = sourceLanguage || detectLanguage(text)

  // Simulate API delay
  await new Promise(resolve => setTimeout(resolve, 500))

  // Mock translation (just add a prefix)
  const translatedText = `[Translated to ${targetLanguage}] ${text}`

  return {
    translatedText,
    sourceLanguage: detectedSource,
  }
}

export async function POST(request: NextRequest) {
  try {
    const body: TranslationRequest = await request.json()

    const { text, targetLanguage, sourceLanguage } = body

    if (!text || !text.trim()) {
      return NextResponse.json(
        { error: 'Text is required' },
        { status: 400 }
      )
    }

    if (!targetLanguage) {
      return NextResponse.json(
        { error: 'Target language is required' },
        { status: 400 }
      )
    }

    if (text.length > 5000) {
      return NextResponse.json(
        { error: 'Text exceeds maximum length of 5000 characters' },
        { status: 400 }
      )
    }

    logger.debug('Translating text', {
      textLength: text.length,
      targetLanguage,
      sourceLanguage,
    })

    const result = await translateText(text, targetLanguage, sourceLanguage)

    logger.info('Translation completed', {
      sourceLanguage: result.sourceLanguage,
      targetLanguage,
    })

    return NextResponse.json({
      translatedText: result.translatedText,
      sourceLanguage: result.sourceLanguage,
      targetLanguage,
    })
  } catch (error) {
    logger.error('Translation API error', error instanceof Error ? error : new Error(String(error)))
    return NextResponse.json(
      { error: error instanceof Error ? error.message : 'Internal server error' },
      { status: 500 }
    )
  }
}

export const runtime = 'edge'
export const dynamic = 'force-dynamic'